requirements.txt
# Generated from `pip freeze` on 2026-01-25
# Python >= 3.10 is required for this project
pyaes==1.6.1
pyasn1==0.6.2
rsa==4.9.1
Telethon==1.42.0
python-dotenv
asyncpg>=0.29.0
groq>=0.9.0
cerebras-cloud-sdk>=1.0.0
google-genai>=0.5.0
requests>=2.30.0

# Web UI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
